requirements.txt
streamlit
pandas
selenium
webdriver-manager
google-generativeai
